[expo-updates][android][4/n] Make a bunch of UpdatesController private
